refact(e2e): Split up boundary-cli helper methods
