+++ ui/admin/tests/e2e/helpers/boundary-cli/session-recordings.js
@@ -1,0 +1,32 @@
+/**
+ * Copyright (c) HashiCorp, Inc.
+ * SPDX-License-Identifier: BUSL-1.1
+ */
+
+const { execSync } = require('child_process');
+
+/**
+ * Waits for a session recording to be available
+ * @param {string} storageBucketId ID of storage bucket that the session recording is associated with
+ * @returns An object representing a session recording
+ */
+export async function waitForSessionRecordingCli(storageBucketId) {
+  let i = 0;
+  let filteredSessionRecording = [];
+  do {
+    i = i + 1;
+    const sessionRecordings = JSON.parse(
+      execSync('boundary session-recordings list --recursive -format json'),
+    );
+    filteredSessionRecording = sessionRecordings.items.filter(
+      (obj) =>
+        obj.storage_bucket_id == storageBucketId && obj.state == 'available',
+    );
+    if (filteredSessionRecording.length > 0) {
+      break;
+    }
+    await new Promise((resolve) => setTimeout(resolve, 1000));
+  } while (i < 5);
+
+  return filteredSessionRecording[0];
+}
