+++ ui/admin/tests/e2e/tests/alias-ent.spec.js
@@ -4,16 +4,18 @@
  */
 
 const { test, expect } = require('@playwright/test');
-import { execSync } from 'child_process';
 import { customAlphabet } from 'nanoid';
 
 const { checkEnv, authenticatedState } = require('../helpers/general');
 const {
   authenticateBoundaryCli,
-  authorizeAlias,
+  authorizeSessionByAliasCli,
   checkBoundaryCli,
   deleteAliasCli,
-  deleteOrgCli,
+  deleteScopeCli,
+  getOrgIdFromNameCli,
+  getProjectIdFromNameCli,
+  getTargetIdFromNameCli,
 } = require('../helpers/boundary-cli');
 const AliasesPage = require('../pages/aliases');
 const CredentialStoresPage = require('../pages/credential-stores');
@@ -89,7 +91,7 @@ test.describe('Aliases (Enterprise)', async () => {
         process.env.E2E_PASSWORD_ADMIN_LOGIN_NAME,
         process.env.E2E_PASSWORD_ADMIN_PASSWORD,
       );
-      await authorizeAlias(alias);
+      await authorizeSessionByAliasCli(alias);
 
       // Clear destination from alias
       await page.getByRole('link', { name: alias }).click();
@@ -108,10 +110,9 @@ test.describe('Aliases (Enterprise)', async () => {
         process.env.E2E_PASSWORD_ADMIN_PASSWORD,
       );
       if (orgName) {
-        const orgs = JSON.parse(execSync('boundary scopes list -format json'));
-        const org = orgs.items.filter((obj) => obj.name == orgName)[0];
-        if (org) {
-          await deleteOrgCli(org.id);
+        const orgId = await getOrgIdFromNameCli(orgName);
+        if (orgId) {
+          await deleteScopeCli(orgId);
         }
       }
       if (alias) {
@@ -158,7 +159,7 @@ test.describe('Aliases (Enterprise)', async () => {
       );
 
       // Connect to target using alias
-      await authorizeAlias(alias);
+      await authorizeSessionByAliasCli(alias);
     } finally {
       await authenticateBoundaryCli(
         process.env.BOUNDARY_ADDR,
@@ -168,10 +169,9 @@ test.describe('Aliases (Enterprise)', async () => {
       );
 
       if (orgName) {
-        const orgs = JSON.parse(execSync('boundary scopes list -format json'));
-        const org = orgs.items.filter((obj) => obj.name == orgName)[0];
-        if (org) {
-          await deleteOrgCli(org.id);
+        const orgId = await getOrgIdFromNameCli(orgName);
+        if (orgId) {
+          await deleteScopeCli(orgId);
         }
         if (alias) {
           await deleteAliasCli(alias);
@@ -183,7 +183,7 @@ test.describe('Aliases (Enterprise)', async () => {
   test('Set up alias from aliases page @ent @aws @docker', async ({ page }) => {
     await page.goto('/');
     const nanoid = customAlphabet('1234567890abcdefghijklmnopqrstuvwxyz', 10);
-    let org;
+    let orgId;
     let alias;
     try {
       const orgsPage = new OrgsPage(page);
@@ -214,24 +214,14 @@ test.describe('Aliases (Enterprise)', async () => {
         process.env.E2E_PASSWORD_ADMIN_LOGIN_NAME,
         process.env.E2E_PASSWORD_ADMIN_PASSWORD,
       );
-
-      const orgs = JSON.parse(execSync('boundary scopes list -format json'));
-      org = orgs.items.filter((obj) => obj.name == orgName)[0];
-      const projects = JSON.parse(
-        execSync('boundary scopes list -format json -scope-id ' + org.id),
-      );
-      const project = projects.items.filter(
-        (obj) => obj.name == projectName,
-      )[0];
-      const targets = JSON.parse(
-        execSync('boundary targets list -format json -scope-id ' + project.id),
-      );
-      const target = targets.items.filter((obj) => obj.name == targetName)[0];
+      orgId = await getOrgIdFromNameCli(orgName);
+      const projectId = await getProjectIdFromNameCli(orgId, projectName);
+      const targetId = await getTargetIdFromNameCli(projectId, targetName);
 
       alias = 'example.alias.' + nanoid();
       const aliasesPage = new AliasesPage(page);
-      await aliasesPage.createAliasForTarget(alias, target.id);
-      await authorizeAlias(alias);
+      await aliasesPage.createAliasForTarget(alias, targetId);
+      await authorizeSessionByAliasCli(alias);
     } finally {
       await authenticateBoundaryCli(
         process.env.BOUNDARY_ADDR,
@@ -240,8 +230,8 @@ test.describe('Aliases (Enterprise)', async () => {
         process.env.E2E_PASSWORD_ADMIN_PASSWORD,
       );
 
-      if (org) {
-        await deleteOrgCli(org.id);
+      if (orgId) {
+        await deleteScopeCli(orgId);
       }
       if (alias) {
         await deleteAliasCli(alias);
