+++ ui/admin/tests/e2e/tests/delete-resources-ent.spec.js
@@ -28,9 +28,9 @@ const {
   createSshTargetCli,
   createVaultCredentialStoreCli,
   createUsernamePasswordCredentialCli,
-  deleteOrgCli,
+  deleteScopeCli,
 } = require('../helpers/boundary-cli');
-const { checkVaultCli } = require('../helpers/vault-cli');
+const { checkVaultCli, getVaultToken } = require('../helpers/vault-cli');
 const AuthMethodsPage = require('../pages/auth-methods');
 const BaseResourcePage = require('../pages/base-resource');
 const WorkersPage = require('../pages/workers');
@@ -95,11 +95,14 @@ test('Verify resources can be deleted (enterprise) @ent @aws', async ({
     let staticHostSetId = await createHostSetCli(staticHostCatalogId);
     let staticCredentialStoreId =
       await createStaticCredentialStoreCli(projectId);
+    const vaultToken = await getVaultToken(
+      boundaryPolicyName,
+      secretPolicyName,
+    );
     let vaultCredentialStoreId = await createVaultCredentialStoreCli(
       projectId,
       process.env.E2E_VAULT_ADDR,
-      secretPolicyName,
-      boundaryPolicyName,
+      vaultToken,
     );
     let usernamePasswordCredentialId =
       await createUsernamePasswordCredentialCli(staticCredentialStoreId);
@@ -175,7 +178,7 @@ test('Verify resources can be deleted (enterprise) @ent @aws', async ({
     orgDeleted = true;
   } finally {
     if (orgId && orgDeleted == false) {
-      await deleteOrgCli(orgId);
+      await deleteScopeCli(orgId);
     }
   }
 });
