+++ ui/admin/tests/e2e/tests/group-role.spec.js
@@ -4,13 +4,13 @@
  */
 
 const { test } = require('@playwright/test');
-const { execSync } = require('child_process');
 
 const { authenticatedState } = require('../helpers/general');
 const {
   authenticateBoundaryCli,
   checkBoundaryCli,
-  deleteOrgCli,
+  deleteScopeCli,
+  getOrgIdFromNameCli,
 } = require('../helpers/boundary-cli');
 const GroupsPage = require('../pages/groups');
 const OrgsPage = require('../pages/orgs');
@@ -44,10 +44,9 @@ test('Verify a new role can be created and associated with a group @ce @ent @aws
       process.env.E2E_PASSWORD_ADMIN_LOGIN_NAME,
       process.env.E2E_PASSWORD_ADMIN_PASSWORD,
     );
-    const orgs = JSON.parse(execSync('boundary scopes list -format json'));
-    const org = orgs.items.filter((obj) => obj.name == orgName)[0];
-    if (org) {
-      await deleteOrgCli(org.id);
+    const orgId = await getOrgIdFromNameCli(orgName);
+    if (orgId) {
+      await deleteScopeCli(orgId);
     }
   }
 });
