+++ ui/admin/tests/e2e/tests/ssh-credential-injection-vault-ent.spec.js
@@ -11,7 +11,10 @@ const {
   authenticateBoundaryCli,
   checkBoundaryCli,
   connectSshToTarget,
-  deleteOrgCli,
+  deleteScopeCli,
+  getOrgIdFromNameCli,
+  getProjectIdFromNameCli,
+  getTargetIdFromNameCli,
 } = require('../helpers/boundary-cli');
 const { checkVaultCli } = require('../helpers/vault-cli');
 const CredentialStoresPage = require('../pages/credential-stores');
@@ -52,7 +55,7 @@ test.beforeEach(async ({ page }) => {
 test('SSH Credential Injection (Vault User & Key Pair) @ent @docker', async ({
   page,
 }) => {
-  let org;
+  let orgId;
   let connect;
   try {
     // Set up vault
@@ -85,22 +88,10 @@ test('SSH Credential Injection (Vault User & Key Pair) @ent @docker', async ({
     // Create org
     const orgsPage = new OrgsPage(page);
     const orgName = await orgsPage.createOrg();
-    await authenticateBoundaryCli(
-      process.env.BOUNDARY_ADDR,
-      process.env.E2E_PASSWORD_AUTH_METHOD_ID,
-      process.env.E2E_PASSWORD_ADMIN_LOGIN_NAME,
-      process.env.E2E_PASSWORD_ADMIN_PASSWORD,
-    );
-    const orgs = JSON.parse(execSync('boundary scopes list -format json'));
-    org = orgs.items.filter((obj) => obj.name == orgName)[0];
 
     // Create project
     const projectsPage = new ProjectsPage(page);
     const projectName = await projectsPage.createProject();
-    const projects = JSON.parse(
-      execSync('boundary scopes list -format json -scope-id ' + org.id),
-    );
-    const project = projects.items.filter((obj) => obj.name == projectName)[0];
 
     // Create target
     const targetsPage = new TargetsPage(page);
@@ -108,10 +99,6 @@ test('SSH Credential Injection (Vault User & Key Pair) @ent @docker', async ({
       process.env.E2E_TARGET_ADDRESS,
       process.env.E2E_TARGET_PORT,
     );
-    const targets = JSON.parse(
-      execSync('boundary targets list -format json -scope-id ' + project.id),
-    );
-    const target = targets.items.filter((obj) => obj.name == targetName)[0];
 
     // Create credentials
     const credentialStoresPage = new CredentialStoresPage(page);
@@ -130,7 +117,16 @@ test('SSH Credential Injection (Vault User & Key Pair) @ent @docker', async ({
     );
 
     // Connect to target
-    connect = await connectSshToTarget(target.id);
+    await authenticateBoundaryCli(
+      process.env.BOUNDARY_ADDR,
+      process.env.E2E_PASSWORD_AUTH_METHOD_ID,
+      process.env.E2E_PASSWORD_ADMIN_LOGIN_NAME,
+      process.env.E2E_PASSWORD_ADMIN_PASSWORD,
+    );
+    orgId = await getOrgIdFromNameCli(orgName);
+    const projectId = await getProjectIdFromNameCli(orgId, projectName);
+    const targetId = await getTargetIdFromNameCli(projectId, targetName);
+    connect = await connectSshToTarget(targetId);
     const sessionsPage = new SessionsPage(page);
     await sessionsPage.waitForSessionToBeVisible(targetName);
     await page
@@ -139,8 +135,8 @@ test('SSH Credential Injection (Vault User & Key Pair) @ent @docker', async ({
       .getByRole('button', { name: 'Cancel' })
       .click();
   } finally {
-    if (org) {
-      await deleteOrgCli(org.id);
+    if (orgId) {
+      await deleteScopeCli(orgId);
     }
     // End `boundary connect` process
     if (connect) {
